Added Add Ingredient Dialog

+++ src/main/java/cz/muni/fi/pv168/project/ui/dialog/RecipeDialog.java
@@ -68,7 +68,7 @@ public class RecipeDialog extends EntityDialog<Recipe> {
         recipe.setName(recipeNameTextField.getText());
         recipe.setDescription(briefDescriptionTextArea.getText());
         recipe.setNumOfServings((int)numberOfServingsSpinner.getValue());
-        recipe.setPreparationTime((int)preparationTimeSlider.getValue());
+        recipe.setPreparationTime(preparationTimeSlider.getValue());
         //recipe.addRecipeCategory((RecipeCategory)recipeCategoryCompoBox.getSelectedItem());
 
         return recipe;
